requirements.txt
google-genai
python-dotenv
streamlit
pandas
numpy
scikit-learn
lightgbm
joblib
